Updated files to include 3 save options
Updated "iterative" in EZFIO.cfg to be an integer that can take the values of 1 (Append), 2 (Overwrite), or 3 (Do not save). 
Updated fci_iterations to be simpler and include the three options.

Also updated the output so that only N_det, energy, and pt2 are output. The user must manipulate from there.